Agrego estilos al navbar y places
