+++ src/pages/Places/index.ts
@@ -1,0 +1,1 @@
+export * from './Places'
