+++ src/utils/dataEscuelas.ts
@@ -1,0 +1,109 @@
+export const escuelas = {
+    "type": "FeatureCollection",
+    "name": "escuelas_integracion",
+    "crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:OGC:1.3:CRS84" } },
+    "features": [
+    { "type": "Feature", "properties": { "id": 1, "Nombre": "Centro Privado de Estimulación Temprana N°18 Santa Rita", "Dirección": "Rivadavia 546, Gualeguaychú. Entre Ríos, Argentina.", "Teléfono": "03446-431173", "Gestión": "Privada", "Descripcion": "Educ. Privada. Ofrece Educación Especial no confesional. Educación Especial: N. Inicial, N. Primaria. Estimulación Temprana, Ayuda a la Integración Escolar.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/Asociaci%C3%B3n-SANTA-RITA-434309569982784", "Email": "-", "lat": -33.00712, "lng": -58.51085 }, "geometry": { "type": "Point", "coordinates": [ -58.510850066838316, -33.007119539993191 ] } },
+    { "type": "Feature", "properties": { "id": 2, "Nombre": "Escuela Integral N°12 Jose Facio", "Dirección": "Andrade 545. ( esquina  3 de Febrero), Gualeguaychú, Entre Ríos.", "Teléfono": "03446-427268", "Gestión": "Pública", "Descripcion": "Educación Especial: Primaria, Estimulación Temprana, Ayuda a la Integración Escolar. Nacida por iniciativa de padres que apostaron a una educación integral.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/escuela.facio ", "Email": "karoclaudia3636@gmail.com.ar", "lat": -33.011695, "lng": -58.51067 }, "geometry": { "type": "Point", "coordinates": [ -58.510670321983788, -33.011695001310926 ] } },
+    { "type": "Feature", "properties": { "id": 3, "Nombre": "Escuela Integral N°2 Francisco Antonio Rizzuto", "Dirección": "Rivadavia N° 831, Gualeguaychú.", "Teléfono": "03446-427233", "Gestión": "Pública", "Descripcion": "Acompaña a personas con discapacidad. Trabaja en la educación integral, junto a más de 120 estudiantes acompañando procesos de aprendizajes.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/EEIRizzuto", "Email": "patriciaebianchihotmail.com", "lat": -33.007177, "lng": -58.515849 }, "geometry": { "type": "Point", "coordinates": [ -58.515848797854105, -33.007177186822403 ] } },
+    { "type": "Feature", "properties": { "id": 4, "Nombre": "Escuela Privada Especial N°28 I.D.E.E.A. y Centro de Dia.", "Dirección": "Bolivar 635, Gualeguaychú. Entre Ríos, Argentina.", "Teléfono": "03446-432253", "Gestión": "Privada", "Descripcion": "Surgió en 1996 con el objetivo de brindar un abordaje multi e interdisciplinario a personas con discapacidad intelectual, física y trastornos asociados.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/profile.php?id=100013300661417", "Email": "institutoideea@gmail.com", "lat": -33.010961, "lng": -58.512192 }, "geometry": { "type": "Point", "coordinates": [ -58.512191510541143, -33.010961206564147 ] } },
+    { "type": "Feature", "properties": { "id": 5, "Nombre": "Escuela Privada Especial N°23 Emmanuel.", "Dirección": "Ramirez N°23 entre calle Urquiza y Luis N. Palma, Gualeguaychú.", "Teléfono": "03446-427274", "Gestión": "Privada", "Descripcion": "Por disposición Nº 136, D.E.P., el 10 de agosto de 1994: Se autoriza el funcionamiento de la Escuela Primaria Especial.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/epei23emanuel", "Email": "escespemanuel@hotmail.com", "lat": -33.00907, "lng": -58.531336 }, "geometry": { "type": "Point", "coordinates": [ -58.531336451448759, -33.009069659880318 ] } },
+    { "type": "Feature", "properties": { "id": 6, "Nombre": "Escuela Integral Privada Nº 12 HORIZONTES ESPECIAL", "Dirección": "Manuel Olloquiegui 171, Larroque. Entre Ríos, Argentina.", "Teléfono": "03446-460249", "Gestión": "Privada", "Descripcion": "La Escuela Horizontes se creó el 8 de agosto de 1977 por iniciativa de la Psp. Ma. del Rosario Cis de Cardoso y la Trabajadora Social Hermana Irma Britez.", "Departamen": "Gualeguaychú", "RRSS": "https:/www.youtube.com/watch?v=8ZJcLrn1YsI", "Email": "escuela_horizontes@hotmail.com.ar", "lat": -33.03144, "lng": -59.006499 }, "geometry": { "type": "Point", "coordinates": [ -59.006498787378419, -33.031440253507043 ] } },
+    { "type": "Feature", "properties": { "id": 7, "Nombre": "Escuela Privada Integral N° 15 Taller Protegido", "Dirección": "Pasteur 75, Gualeguaych. Entre Ríos, Argentina", "Teléfono": "03446-428621", "Gestión": "Privada", "Descripcion": "La escuela surge en 1981, ante la carencia de ofertas educativas para jóvenes y adultos de esta comunidad. Frente a esto se conforma la Asociación Emanuel", "Departamen": "Gualeguaychú", "RRSS": "https:/www.facebook.com/EPEI-N15-Taller-Protegido-Emanuel-1483976358497990/", "Email": "talleremanuel81@gmail.com", "lat": -33.010586, "lng": -58.531821 }, "geometry": { "type": "Point", "coordinates": [ -58.531821178593461, -33.01058568863489 ] } },
+    { "type": "Feature", "properties": { "id": 8, "Nombre": "Escuela Especial N° 13 Arco Iris", "Dirección": "Pbro. Francisco Kaul 235, Urdinarrain. Entre Ríos, Argentina.", "Teléfono": "03446-480585", "Gestión": "Privada", "Descripcion": "En 1983 se coloca la piedra fundamental del edificio actual, y en 1984 comienza a funcionar la escuela en dicho lugar, sito en N.L. Além Y F. Kaúl.", "Departamen": "Gualeguaychú", "RRSS": "http:/www.arcoirisintegral.org.ar/contacto/contacto.php", "Email": "arcoiris@urdi.com.ar", "lat": -32.681246, "lng": -58.88881 }, "geometry": { "type": "Point", "coordinates": [ -58.888810306896772, -32.681245945047444 ] } },
+    { "type": "Feature", "properties": { "id": 9, "Nombre": "Escuela de Integral N°4 Mayor Pastor", "Dirección": "Federación Urquiza y Federación, Rosario Del Tala, Entre Ríos.", "Teléfono": "Federación, Rosario Del Tala, Entre Ríos.Teléfono:", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles estimulación temprana, inicial y primario.", "Departamen": "Tala", "RRSS": "-", "Email": "1801420-00@mec.gob.ar", "lat": -32.301298, "lng": -59.134394 }, "geometry": { "type": "Point", "coordinates": [ -59.134394277728155, -32.301297768815637 ] } },
+    { "type": "Feature", "properties": { "id": 10, "Nombre": "Escuela Especial N°28 Eva Perón", "Dirección": "Osvaldo Magnasco S/N, Macia.  Entre Rios, Argentina", "Teléfono": "03445-461574", "Gestión": "Pública", "Descripcion": "En Eva Peròn Especial 28 puede cursar:Educación especial, Educación especial primaria,Talleres de formación integral", "Departamen": "Tala", "RRSS": "-", "Email": "escuela_especial_28macia@hotmail.com", "lat": -32.173676, "lng": -59.404854 }, "geometry": { "type": "Point", "coordinates": [ -59.404854466864386, -32.173675806081889 ] } },
+    { "type": "Feature", "properties": { "id": 11, "Nombre": "Escuela Nina N°1 Carlos Maria Alvear", "Dirección": "Maipú 1552,  Nogoya. Entre Ríos, Argentina.", "Teléfono": "03435-421210", "Gestión": "Pública", "Descripcion": "Según datos obtenidos, el edificio se inauguró el 31 de mayo de 1910 llamándose Escuela Superior Mixta Nº 1 Carlos María de Alvear", "Departamen": "Nogoyá", "RRSS": "https:/www.facebook.com/escuelaalvearnogoya", "Email": "escuela_alverar_nogoya@hotmail.com", "lat": -32.399252, "lng": -59.815755 }, "geometry": { "type": "Point", "coordinates": [ -59.815754519665688, -32.399251870215821 ] } },
+    { "type": "Feature", "properties": { "id": 12, "Nombre": "C.I.C Nogoyá: Centro Integrador Comunitario", "Dirección": "Federación y Marconi", "Teléfono": "03435 - 427452", "Gestión": "Pública", "Descripcion": "Destinado a desarrollar tareas vinculadas a la salud y desarrollo social en la ciudad.", "Departamen": "Nogoyá", "RRSS": "https:/www.facebook.com/CIC-y-Centros-de-Salud-Gobierno-de-Nogoy%C3%A1-761464317319976", "Email": "cicnogoya@argentina.com.ar", "lat": -32.39139, "lng": -59.79973 }, "geometry": { "type": "Point", "coordinates": [ -59.799729754030515, -32.391389922068697 ] } },
+    { "type": "Feature", "properties": { "id": 13, "Nombre": "Escuela Privada Integral N°8 El Arca de Noé", "Dirección": "Quiroga y Taboada 1075, Nogoyá. Entre Ríos, Argentina.", "Teléfono": "03435 42-1883", "Gestión": "Privada", "Descripcion": "La institución ofrece enseñanza especial en los niveles estimulación temprana, inicial y primario. EL 17 de abril del 2021 conmemorò el 47º aniversario de la creación de la entidad.", "Departamen": "Nogoyá", "RRSS": "https:/www.facebook.com/pages/Escuela%20Privada%20De%20Educaci%C3%B3n%20Integral%20N8%20El%20Arca%20De%20No%C3%A9/1835656790005609/", "Email": "esc_esp_elarcadenoe@hotmail.com", "lat": -32.398636, "lng": -59.789625 }, "geometry": { "type": "Point", "coordinates": [ -59.789624890524671, -32.398636332756858 ] } },
+    { "type": "Feature", "properties": { "id": 14, "Nombre": "Centro Educativo Terapeutico N°3", "Dirección": "Roque Sáenz Peña, Lucas González, Entre Ríos", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "Brinda un abordaje pedagógico terapéutico integral a niños y jóvenes con discapacidad que impacte positivamente en la mejora de sus condiciones de vida y de la de sus familias.", "Departamen": "Nogoyá", "RRSS": "https:/www.facebook.com/ceitmilugar", "Email": "ceit_lg@hotmail.com.ar", "lat": -32.386106, "lng": -59.535051 }, "geometry": { "type": "Point", "coordinates": [ -59.535050849056205, -32.386106085767409 ] } },
+    { "type": "Feature", "properties": { "id": 15, "Nombre": "Escuela Integral N°15 San Francisco Javier", "Dirección": "Airaldi 601, Diamante, Entre Ríos.", "Teléfono": "0343 498-4171", "Gestión": "Pública", "Descripcion": "En la escuela San Francisco Javier de Diamante se busca y se consigue la integración posterior de los estudiantes en cualquier otro  establecimiento educativo.", "Departamen": "Diamante", "RRSS": "-", "Email": "-", "lat": -32.07269, "lng": -60.630702 }, "geometry": { "type": "Point", "coordinates": [ -60.630702154939215, -32.072690478180867 ] } },
+    { "type": "Feature", "properties": { "id": 16, "Nombre": "Escuela Integral N° 21 Portal de la Esperanza", "Dirección": "Av. De los Pioneros S/N, Villa Libertador San Martín Diamante Entre Ríos", "Teléfono": "0343-4911327", "Gestión": "Privada", "Descripcion": "La institución ofrece enseñanza especial en los niveles inicial y primario.", "Departamen": "Diamante", "RRSS": "https:/www.facebook.com/Escuela-Privada-de-Educaci%C3%B3n-Integral-N%C2%BA-21-Portal-de-la-Esperanza-102913199795046", "Email": "portaldelaesperanza@yahoo.com.ar", "lat": -32.062915, "lng": -60.449655 }, "geometry": { "type": "Point", "coordinates": [ -60.449654945127072, -32.062915297779128 ] } },
+    { "type": "Feature", "properties": { "id": 17, "Nombre": "Escuela Especial N° 44 Enrique de Vedia", "Dirección": "Sanatorio Fidanza, Colonia Ensayo. Entre Ríos, Argentina.", "Teléfono": "4978120 INT 117", "Gestión": "Pública", "Descripcion": "-", "Departamen": "-", "RRSS": "-", "Email": "-", "lat": -31.855447, "lng": -60.617397 }, "geometry": { "type": "Point", "coordinates": [ -60.617397174741576, -31.85544737489592 ] } },
+    { "type": "Feature", "properties": { "id": 18, "Nombre": "Escuela Privada Especial N°5Alborada", "Dirección": "Calle Irigoyen, General Ramirez, Entre Ríos", "Teléfono": "0343-4901488", "Gestión": "Privada", "Descripcion": "Institución Privada con las modalidades de Integración, Estimulación Temprana en los Niveles Inicial, Primario con Pre Talleres y Talleres.", "Departamen": "Diamante", "RRSS": "-", "Email": "especialalborada@hotmail.com", "lat": -32.177334, "lng": -60.203707 }, "geometry": { "type": "Point", "coordinates": [ -60.203706691892478, -32.177334213701045 ] } },
+    { "type": "Feature", "properties": { "id": 19, "Nombre": "Escuela Integral Privada N° 4 DINAD", "Dirección": "Hipolito Irigoyen 420, Victoria. Entre Ríos, Argentina.", "Teléfono": "03436-421604", "Gestión": "Privada", "Descripcion": "-", "Departamen": "-", "RRSS": "https:/es-la.facebook.com/DINAD-499606336732438/", "Email": "dinad@live.com.ar", "lat": -32.625439, "lng": -60.154475 }, "geometry": { "type": "Point", "coordinates": [ -60.154474970709302, -32.62543882787373 ] } },
+    { "type": "Feature", "properties": { "id": 20, "Nombre": "CIC - Centro Integrador Comunitario Victoria", "Dirección": "Cochabamba, entre Paso y Arenales.", "Teléfono": "03436 - 424584", "Gestión": "Pública", "Descripcion": "Enfermería:Guardia en ambos turnos.Extracción de puntos.Control de glucemia TA.Más de 8000 historias clínicas.Control de niños sanos, talla, peso y perímetro encefálico.", "Departamen": "Victoria", "RRSS": "-", "Email": "informescic@arnetbiz.com.ar", "lat": -32.608263, "lng": -60.154142 }, "geometry": { "type": "Point", "coordinates": [ -60.154142188805451, -32.608263483219197 ] } },
+    { "type": "Feature", "properties": { "id": 21, "Nombre": "APAFADI. Servicio para personas con discapacidad.", "Dirección": "Vélez Sársfield 387, Victoria. Entre Ríos, Argentina.", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "Servicios para personas con discapacidad, profesores en educación especial, profesores de educación física, escuelas jardines de infantes, profesionales de la salud y más.", "Departamen": "Victoria", "RRSS": "https:/es-la.facebook.com/APAFADI/", "Email": "apafadi_2000@outlook.es", "lat": -32.619767, "lng": -60.161241 }, "geometry": { "type": "Point", "coordinates": [ -60.161241107353945, -32.619767208124657 ] } },
+    { "type": "Feature", "properties": { "id": 22, "Nombre": "Escuela Integral N°43 Provincia de Buenos Aires", "Dirección": "San Martín 276, Victoria. Entre Rios, Argentina.", "Teléfono": "03436-422198", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Victoria", "RRSS": "https:/es-la.facebook.com/pages/category/Elementary-School/Escuela-N%C2%BA43-Provincia-de-Buenos-Aires-1416145195304312/", "Email": "-", "lat": -32.617753, "lng": -60.160954 }, "geometry": { "type": "Point", "coordinates": [ -60.16095402888795, -32.617752836845213 ] } },
+    { "type": "Feature", "properties": { "id": 23, "Nombre": "Escuela Privada Integral N°19 Lucecita", "Dirección": "Salta 82, Gualeguay. Entre Ríos, Argentina.", "Teléfono": "03444-422825", "Gestión": "Privada", "Descripcion": "La Escuela de Educación Integral, fue creada el 18 de septiembre de 1967 por un grupo de Padres que aceptó el desafío de una Institución que pudiera recibir a sus hijos.", "Departamen": "Gualeguay", "RRSS": "https:/www.facebook.com/municipalidaddegualeguay/posts/2703366066657920/", "Email": "LUCESITAS@ARNETBIZ.COM.AR", "lat": -33.148502, "lng": -59.316404 }, "geometry": { "type": "Point", "coordinates": [ -59.316403501210203, -33.148502078467168 ] } },
+    { "type": "Feature", "properties": { "id": 24, "Nombre": "Escuela Especial N° 19 Leon Martinelli", "Dirección": "Leandro N. Alem 320, Gualeguay. Entre Ríos, Argentina", "Teléfono": "03444-424586", "Gestión": "Pública", "Descripcion": "-", "Departamen": "-", "RRSS": "-", "Email": "escuelaespecial19@hotmail.com.ar", "lat": -33.146818, "lng": -59.318734 }, "geometry": { "type": "Point", "coordinates": [ -59.318733739311334, -33.146817615275388 ] } },
+    { "type": "Feature", "properties": { "id": 25, "Nombre": "Escuela de Educación Integral Nº 32 Alcancías de ilusiones", "Dirección": "López y planes, y Martín Fierro (arco), Pueblo Ibicuy.", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "Creada en 1991, cuenta con el servicio de integración/inclusión dentro de sus especialidades, con tareas de alfabetización, orientación vocacional, talleres de formación laboral en auxiliar panadero.", "Departamen": "Ibicuy", "RRSS": "-", "Email": "especial.32@hotmail.com", "lat": -33.746874, "lng": -59.145789 }, "geometry": { "type": "Point", "coordinates": [ -59.145789190030072, -33.746873993402446 ] } },
+    { "type": "Feature", "properties": { "id": 26, "Nombre": "Centro de Educación Integral N°8 Nuestra Señora del Rosario (CEIT N°8).", "Dirección": "Villa Paranacito. Domicilio: Paraná s/n cerro poblacional. Ibicuy. Entre Ríos, Argentina.", "Teléfono": "03446-495310", "Gestión": "Pública", "Descripcion": "-", "Departamen": "-", "RRSS": "-", "Email": "susanaroude@hotmail.com", "lat": -33.714181, "lng": -58.65961 }, "geometry": { "type": "Point", "coordinates": [ -58.659610487894277, -33.714181297019138 ] } },
+    { "type": "Feature", "properties": { "id": 27, "Nombre": "Escuela Especial Nº 2 Rayo de Luz", "Dirección": "9 de julio 5610", "Teléfono": "03447-421474", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles maternal, inicial y primario.", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/groups/581478181903584", "Email": "esc.esp.rayodeluz@hotmail.com", "lat": -32.218567, "lng": -58.144859 }, "geometry": { "type": "Point", "coordinates": [ -58.144858945536775, -32.218567298263082 ] } },
+    { "type": "Feature", "properties": { "id": 28, "Nombre": "Escuela Privada de Educación Integral Nº 25 Adcadis", "Dirección": "Boulevard Ferrari 38, Colón. Entre Ríos, Argentina.", "Teléfono": "03447 42-3546", "Gestión": "Privada", "Descripcion": "Formación Profesional: Auxiliar en Panadería (Auxiliar Panadero / Panificación) y Auxiliar en Vivero y Parquización (Auxiliar de Jardinería y/o de Vivero)", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/ADCADis-116243829029995", "Email": "hola@adcadis.org aradcadis@yahoo.com.ar", "lat": -32.230199, "lng": -58.13935 }, "geometry": { "type": "Point", "coordinates": [ -58.139350333465991, -32.230199430039988 ] } },
+    { "type": "Feature", "properties": { "id": 29, "Nombre": "Escuela de Educación Integral N° 29 Centenario", "Dirección": "Av. Mitre 685, Villa Elisa. Entre Ríos, Argentina.", "Teléfono": "0221 4643426", "Gestión": "Pública", "Descripcion": "Tipo de Institución: Formación Profesional", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/watch/?v=747513156027209", "Email": "escuelaespecial501berisso@gmail.com", "lat": -32.16858, "lng": -58.407188 }, "geometry": { "type": "Point", "coordinates": [ -58.40718844795785, -32.168580080873475 ] } },
+    { "type": "Feature", "properties": { "id": 30, "Nombre": "Escuela de Educación Integral N° 30 Dr. Augusto Vanerio", "Dirección": "Malvinas Argentinas 911, San José. Entre Ríos, Argentina.", "Teléfono": "3447470416", "Gestión": "Pública", "Descripcion": "Educación Especial, Educación Especial Nivel Inicial, Educación Especial Primaria, Estimulación Temprana.", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/escuela.30", "Email": "escuelaespecial30@hotmail.com", "lat": -32.213133, "lng": -58.214582 }, "geometry": { "type": "Point", "coordinates": [ -58.214581508493893, -32.213132817656081 ] } },
+    { "type": "Feature", "properties": { "id": 31, "Nombre": "Centro Educativo Terapéutico El Solar", "Dirección": "Paysandú 267, Colón, Entre Ríos", "Teléfono": "03447 42-4051", "Gestión": "Privada", "Descripcion": "Escuela Privada de Educación Integral (EPEI) para niños, jóvenes y adultos Y Centro Educativo Terapéutico (C.E.T.)  Intervención Temprana", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/centroelsolarcolon", "Email": "info@elsolarcolon.com.ar", "lat": -32.23325, "lng": -58.133339 }, "geometry": { "type": "Point", "coordinates": [ -58.133338862409794, -32.233250486757257 ] } },
+    { "type": "Feature", "properties": { "id": 32, "Nombre": "Escuela Privada Especial N° 24  El Solar", "Dirección": "San Martín 91, Colón. Entre Ríos, Argentina.", "Teléfono": "03447-421994", "Gestión": "Privada", "Descripcion": "Los talleres que se brindan actualmente son: Vitro fusión, Carpintería, Telar, Computación, Manualidades, Mantenimiento e Imagen.", "Departamen": "Colón", "RRSS": "-", "Email": "cetelsolar@arnet.com.ar", "lat": -32.218306, "lng": -58.136061 }, "geometry": { "type": "Point", "coordinates": [ -58.136061027540528, -32.218306414636572 ] } },
+    { "type": "Feature", "properties": { "id": 33, "Nombre": "No me Olvides – Asociación Síndrome de Down Colón", "Dirección": "Alberdi 314, Colón. Entre Ríos, Argentina.", "Teléfono": "03447-15459348", "Gestión": "Pública", "Descripcion": "La Asociación Síndrome de Down Colón No me olvides, nuclea a familiares y amigos de personas con síndrome de Down sobre la base de entender que comparten una misma realidad.", "Departamen": "Colón", "RRSS": "https:/m.facebook.com/pages/category/Nonprofit-Organization/Asociaci%C3%B3n-S%C3%ADndrome-de-Down-Col%C3%B3n-No-me-olvides-111508975563393/?locale2=es_LA", "Email": "-", "lat": -32.218259, "lng": -58.142599 }, "geometry": { "type": "Point", "coordinates": [ -58.142599412077999, -32.218259417705866 ] } },
+    { "type": "Feature", "properties": { "id": 34, "Nombre": "Fundación Le Ralos", "Dirección": "Paysandú 267 de Colón, Entre Ríos.", "Teléfono": "03447/421994/ 03447-15413240/03447-424051", "Gestión": "Pública", "Descripcion": "ONG sin fines de lucro cuyo objeto es fomentar la integración, educación, rehabilitación y reinserción de todas aquellas personas que padezcan una alteración funcional...", "Departamen": "Colon", "RRSS": "https:/www.facebook.com/leraloscolon/", "Email": "leralos@elsolarcolon.com.ar", "lat": -32.21467, "lng": -58.144935 }, "geometry": { "type": "Point", "coordinates": [ -58.144934580096368, -32.214669788526322 ] } },
+    { "type": "Feature", "properties": { "id": 35, "Nombre": "Centro de Rehabilitación Integral y Centro de Día Santa Ana", "Dirección": "Sabatier 224, Colón. Entre Ríos, Argentina.", "Teléfono": "03447 42-7339", "Gestión": "Privada", "Descripcion": "Centro de Rehabilitación Integral y Centro de Día para la atención de personas con discapacidad durante la jornada diurna; brindando atención a personas con o sin deterioro funcional...", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/santaanacolon/", "Email": "info@santaanacolon.com.ar", "lat": -32.209873, "lng": -58.148406 }, "geometry": { "type": "Point", "coordinates": [ -58.148405562212467, -32.209872771034767 ] } },
+    { "type": "Feature", "properties": { "id": 36, "Nombre": "Grupo T.E.A. Colón", "Dirección": "Güemes 435, Colón. Entre Ríos, Argentina.", "Teléfono": "03447-1550310 03447 - 15 451357", "Gestión": "Privada", "Descripcion": "Grupo de padres, madres y profesionales organizados por la implementación de la ley 27043 en Entre Ríos y la detección temprana de TEA", "Departamen": "Colón", "RRSS": "https:/www.facebook.com/Grupo-TEA-Col%C3%B3n-1917773498282456/", "Email": "Lilibidal@gmail.com", "lat": -32.230534, "lng": -58.137045 }, "geometry": { "type": "Point", "coordinates": [ -58.137045014645111, -32.230533500397677 ] } },
+    { "type": "Feature", "properties": { "id": 37, "Nombre": "Escuela Integral Nº 1 Concordia", "Dirección": "Estrada 552 y Lamadrid, Concordia.  Entre Ríos, Argentina.", "Teléfono": "0345 4215562", "Gestión": "Pública", "Descripcion": "La institución pública cuenta con los niveles de educación temprana, inicial y primaria.", "Departamen": "Concordia", "RRSS": "-", "Email": "-", "lat": -31.394045, "lng": -58.00642 }, "geometry": { "type": "Point", "coordinates": [ -58.00641975183818, -31.394044963297617 ] } },
+    { "type": "Feature", "properties": { "id": 38, "Nombre": "Escuela Integral N° 2 Azahares", "Dirección": "Perú 1325, Concordia, Entre Ríos.", "Teléfono": "03445 461574", "Gestión": "Pública", "Descripcion": "Auxiliar de Huerta y Jardinería, Auxiliar de Marroquinería. Auxiliar de Panadería, Auxiliar en Actividades Agropecuarias, Auxiliar en Carpintería, Ayudante de Cocina", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/laintegral2azahares", "Email": "escuelaintegral28@hotmail.com", "lat": -31.383037, "lng": -58.025854 }, "geometry": { "type": "Point", "coordinates": [ -58.02585360473585, -31.383037201522971 ] } },
+    { "type": "Feature", "properties": { "id": 39, "Nombre": "Escuela Integral N°25 María Ana Mac Cotter de Madrazo", "Dirección": "Eva Perón y De los Viñedos, Concordia, Entre Ríos.", "Teléfono": "0345 - 4210762", "Gestión": "Pública", "Descripcion": "Formación Profesional:Artesano en calado y pintura sobre madera (Artesanías en Madera)Auxiliar Panadero (Auxiliar Panadero / Panificación)", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/Escuela-de-Educaci%C3%B3n-Integral-25-Mar%C3%ADa-Ana-Mac-Cotter-de-Madrazzo-255282007841382", "Email": "escuelaintegral25@gmail.com", "lat": -31.359045, "lng": -58.009231 }, "geometry": { "type": "Point", "coordinates": [ -58.009230522612661, -31.359045361275758 ] } },
+    { "type": "Feature", "properties": { "id": 40, "Nombre": "Escuela Integral Privada N° 17 Arid", "Dirección": "Urquiza N°75, Concordia, Entre Ríos.", "Teléfono": "0345 – 4229462", "Gestión": "Privada", "Descripcion": "Organización ciclada infantil, organización no ciclada infantil, organización no ciclada jóvenes y adultos, orientación vocacional ocupacional, capacitación laboral", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/Escuela-Integral-Arid-150113282232929", "Email": "escuelaincusiva@hotmail.com.ar", "lat": -31.407127, "lng": -58.020033 }, "geometry": { "type": "Point", "coordinates": [ -58.020032741112104, -31.407127299562497 ] } },
+    { "type": "Feature", "properties": { "id": 41, "Nombre": "Centro Privado Educativo Integral N°20 CEREM", "Dirección": "Av. Robinson 50, Concordia, Entre Ríos.", "Teléfono": "0345-421414", "Gestión": "Privada", "Descripcion": "Características: La institución ofrece enseñanza especial en los niveles inicial y nivel primario.", "Departamen": "Concordia", "RRSS": "-", "Email": "cerem@concordia.com.ar", "lat": -31.40247, "lng": -58.012559 }, "geometry": { "type": "Point", "coordinates": [ -58.012558545812553, -31.40247039287258 ] } },
+    { "type": "Feature", "properties": { "id": 42, "Nombre": "Escuela Privada de Educación Integral N°29 Semillas de Esperanza", "Dirección": "Moreno N° 596, Concordia, Entre Ríos.", "Teléfono": "0345 602-740", "Gestión": "Privada", "Descripcion": "Los alumnos reciben educación pedagógica y terapéutica y, además, realizan talleres de dulces, panadería y artesanías.", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/escuelasemilladeesperanza", "Email": "semilladeesperanza@outlook.es", "lat": -31.399661, "lng": -58.006462 }, "geometry": { "type": "Point", "coordinates": [ -58.006461955165904, -31.399660520060916 ] } },
+    { "type": "Feature", "properties": { "id": 43, "Nombre": "Escuela Especial N° 10 Benito Garat", "Dirección": "Urdinarrain 752 Pompeya Sur, Concordia, Entre Ríos.", "Teléfono": "0345-4210822", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles inicial y primario.", "Departamen": "Concordia", "RRSS": "-", "Email": "-", "lat": -31.39226, "lng": -58.033809 }, "geometry": { "type": "Point", "coordinates": [ -58.03380904069153, -31.392260466104268 ] } },
+    { "type": "Feature", "properties": { "id": 44, "Nombre": "Manos a la obra", "Dirección": "Chabrillon 1280 - Concordia - Entre Ríos", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "Grupo recreativo para adolescentes, jóvenes y adultos con discapacidad intelectual", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/Fundacion-Manos-a-La-Obra-193296997382734/", "Email": "laurascattini@gmail.com", "lat": -31.384549, "lng": -57.987354 }, "geometry": { "type": "Point", "coordinates": [ -57.987354114273693, -31.384548500893775 ] } },
+    { "type": "Feature", "properties": { "id": 45, "Nombre": "Asociación de Discapacitados Motores", "Dirección": "1 de Mayo 357, Concordia, Entre Ríos,", "Teléfono": "0345 528-4007", "Gestión": "Privada", "Descripcion": "Características: La institución trabaja con personas con discapacidad, realizando actividades deportivas.", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/ASOCIACION.DISCAPACITADOS.MOTORES.DE.CONCORDIA/posts/asociacion-discapacitados-motores-de-concordia-basket/330112633759121/", "Email": "pino.1965@hotmail.com", "lat": -31.398843, "lng": -58.01144 }, "geometry": { "type": "Point", "coordinates": [ -58.011439687995818, -31.398842990462246 ] } },
+    { "type": "Feature", "properties": { "id": 46, "Nombre": "Asociación Ciegos Concordienses", "Dirección": "Pte. A. Illia y Sarmiento, Concordia, Entre Ríos.", "Teléfono": "0345 421 0636", "Gestión": "Privada", "Descripcion": "Características: La institución realiza diversas actividades y talleres para personas con discapacidad, entre ellas encuentros corales.", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/acicon.acicon", "Email": "-", "lat": -31.374365, "lng": -58.015863 }, "geometry": { "type": "Point", "coordinates": [ -58.015862767115706, -31.374364599909963 ] } },
+    { "type": "Feature", "properties": { "id": 47, "Nombre": "Fundación Ecuestre Los Ceibos", "Dirección": "Av. Eva Perón esq. Poenitz, Concordia, Entre Ríos.", "Teléfono": "(0345) 425 - 2023", "Gestión": "Privada", "Descripcion": "Características: Ofrecen educación especial y talleres de equinoterapia para estimular a niños y niñas con discapacidad.", "Departamen": "Concordia", "RRSS": "http:/www.actiweb.es/losceibos/", "Email": "fundlosceibos@hotmail.com", "lat": -31.370649, "lng": -58.011094 }, "geometry": { "type": "Point", "coordinates": [ -58.011093549841519, -31.370649347041077 ] } },
+    { "type": "Feature", "properties": { "id": 48, "Nombre": "TEA Padres Concordia", "Dirección": "Fernando Victorin 1638, Concordia. Entre Ríos, Argentina.", "Teléfono": "0345 407-1449", "Gestión": "Privada", "Descripcion": "Organizaión No Gubernamental", "Departamen": "Concordia", "RRSS": "https:/www.facebook.com/TEA-Padres-Concordia-415568945202843", "Email": "asociac.tea.concordia@gmail.com", "lat": -31.379575, "lng": -58.008237 }, "geometry": { "type": "Point", "coordinates": [ -58.008236627215638, -31.379575416939137 ] } },
+    { "type": "Feature", "properties": { "id": 49, "Nombre": "Asociación Concordiense de Padres y Amigos del Discapacitado (a.C.o.P.a.C.)", "Dirección": "Mitre 596, Colonia Ayui. Entre Rios, Argentina", "Teléfono": "03454 22 3757 - 03454 22 3757", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Concordia", "RRSS": "-", "Email": "-", "lat": -31.398829, "lng": -58.006353 }, "geometry": { "type": "Point", "coordinates": [ -58.006353105463106, -31.39882905921878 ] } },
+    { "type": "Feature", "properties": { "id": 50, "Nombre": "Escuela Integral N° 9 Dr. Luis Agote", "Dirección": "Francisco Ramírez 69, General Galarza, San Salvador.", "Teléfono": "0345-4910077", "Gestión": "Pública", "Descripcion": "Tipo de Institución: Educación Integral y Formación ProfesionalAuxiliar Huertero/Jardinero (Jardinería)Auxiliar Panadero (Auxiliar Panadero / Panificación)", "Departamen": "San Salvador", "RRSS": "https:/escuelaintegraln9.blogspot.com/", "Email": "-", "lat": -32.721757, "lng": -59.390414 }, "geometry": { "type": "Point", "coordinates": [ -59.390413838743399, -32.721757309038964 ] } },
+    { "type": "Feature", "properties": { "id": 51, "Nombre": "Asociación Padres del Discapacitado de San Salvador", "Dirección": "Mitre 905, San Salvador.  Entre Ríos, Argentina.", "Teléfono": "(03456) 420806", "Gestión": "Privada", "Descripcion": "Busca integrar a las personas con discapacidad, brindandole,tanto a ellos cmo a sus familias las herramientas para alcanzar una mayor participacion", "Departamen": "San Salvador", "RRSS": "https:/www.facebook.com/Asociacion-de-Padres-del-Discapacitado-de-San-Salvador-325692824222991/", "Email": "-", "lat": -31.625487, "lng": -58.514589 }, "geometry": { "type": "Point", "coordinates": [ -58.514589346240072, -31.625487344455838 ] } },
+    { "type": "Feature", "properties": { "id": 52, "Nombre": "Escuela Especial Nº 16 Arturo Illia", "Dirección": "Dr. Chessini 51, Federación. Entre Ríos, Argentina.", "Teléfono": "358 4672954 / 03456-481341", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles inicial, primario y talleres de formación integral.", "Departamen": "Federación", "RRSS": "-", "Email": "escuelaespecial16@federacionnet.com.ar", "lat": -30.986621, "lng": -57.923685 }, "geometry": { "type": "Point", "coordinates": [ -57.923685278843287, -30.986620769989369 ] } },
+    { "type": "Feature", "properties": { "id": 53, "Nombre": "Escuela de Educacion Integral N° 8 Victoria de Fissore", "Dirección": "Repetto 2945 La Tablada, Chajarí, Entre Ríos.", "Teléfono": "Repetto 2945 La Tablada, Chajarí, Entre Ríos.", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles maternal, inicial, primario y secundario.", "Departamen": "Federación", "RRSS": "-", "Email": "ee8_fissore@hotmail.com", "lat": -30.755818, "lng": -57.987719 }, "geometry": { "type": "Point", "coordinates": [ -57.987719060940236, -30.755817673440507 ] } },
+    { "type": "Feature", "properties": { "id": 54, "Nombre": "Centro Privado. Ntra. Sra. De La Luz Especial N° 22", "Dirección": "Mitre 905, Chajarí. Entre Ríos, Argentina.", "Teléfono": "03456-420806", "Gestión": "Privada", "Descripcion": "Estudios que se cursan: Educación especialEducación especial tempranaJardín de infantes de educación especialEducación especial primariaTalleres de formación integral", "Departamen": "Federación", "RRSS": "-", "Email": "escuela.22@hotmail.com", "lat": -30.763675, "lng": -57.982423 }, "geometry": { "type": "Point", "coordinates": [ -57.982422844392246, -30.763675298290131 ] } },
+    { "type": "Feature", "properties": { "id": 55, "Nombre": "Centro Educativo Integral Y Terapéutico N°5", "Dirección": "Sarmiento y Alvarado – San Jaime de La Frontera, Entre Ríos", "Teléfono": "02224 422741", "Gestión": "Pública", "Descripcion": "Auxiliar de Huerta y Vivero (Auxiliar de Jardinería y/o de Vivero).Auxiliar Taller Reciclado (Capacitación en Reciclado de Papeles, Plásticos y/o Metales).", "Departamen": "Federación", "RRSS": "-", "Email": "cei-sanjaime@hotmail.com", "lat": -30.336958, "lng": -58.304437 }, "geometry": { "type": "Point", "coordinates": [ -58.304436883671379, -30.336958018968659 ] } },
+    { "type": "Feature", "properties": { "id": 56, "Nombre": "ASDCU Asociación Síndrome de Down de Concepción del Uruguay", "Dirección": "Dr. J. Lacava 152, Concepción del Uruguay, Entre Ríos", "Teléfono": "424-743 / 423-101", "Gestión": "Privada", "Descripcion": "La entidad se encuentra formada por padres de hijos con Síndrome de Down. Buscan una mejor calidad de vida de los niños integrantes de la asociación.", "Departamen": "Uruguay", "RRSS": "https:/www.facebook.com/groups/251881046180/about", "Email": "-", "lat": -32.486502, "lng": -58.252733 }, "geometry": { "type": "Point", "coordinates": [ -58.252733088988215, -32.486501826778657 ] } },
+    { "type": "Feature", "properties": { "id": 57, "Nombre": "Escuela Especial N°18 Dr. Juan Alberto Marco", "Dirección": "Contador Saúl Rubinsky 1167 Gral. Mosconi, Concepción del Uruguay", "Teléfono": "(03442) - 427229", "Gestión": "Pública", "Descripcion": "Estudios que se cursan:Educación EspecialEducación TempranaJardín de infantesPrimariaTalleres de formación integral", "Departamen": "Uruguay", "RRSS": "https:/www.facebook.com/pg/escuela.juanamarco/posts/", "Email": "escuelaintegral18@gmail.com", "lat": -32.471144, "lng": -58.244356 }, "geometry": { "type": "Point", "coordinates": [ -58.244355878970822, -32.471144331530468 ] } },
+    { "type": "Feature", "properties": { "id": 58, "Nombre": "Escuela de Educación Integral Nº 11 Pedro Ponce de León", "Dirección": "Juan Perón 749, Concepción del Uruguay", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Uruguay", "RRSS": "-", "Email": "-", "lat": -32.495366, "lng": -58.229605 }, "geometry": { "type": "Point", "coordinates": [ -58.229604855039611, -32.495365653510206 ] } },
+    { "type": "Feature", "properties": { "id": 59, "Nombre": "Escuela Integral N°10 Ovidio Decroly", "Dirección": "Rogelio Gómez 155, Basavilbaso, Entre Ríos.", "Teléfono": "03445 - 481033", "Gestión": "Pública", "Descripcion": "Propuestas Educativas: Ayudante de Costura, Blanqueria y Bolsas (Ayudante de Costurera); Ayudante de Envasado de Dulces y Conservas", "Departamen": "Uruguay", "RRSS": "https:/www.facebook.com/Escuela-Integral-N%C2%BA10-Ovidio-Decroly-100392181843437/about/?ref=page_internal", "Email": "ovidiodecroly.esc.int.10@gmail.com", "lat": -32.372822, "lng": -58.874466 }, "geometry": { "type": "Point", "coordinates": [ -58.874466439992631, -32.372822322126382 ] } },
+    { "type": "Feature", "properties": { "id": 60, "Nombre": "Escuela Privada Integral N°1 Surcos de Esperanza", "Dirección": "Fray Santa María De Oro N° 548, Concepción del Uruguay, Entre Ríos.", "Teléfono": "(03442) - 427695", "Gestión": "Privada", "Descripcion": "Propuestas educativas: Educación Especial, Nivel Inicial, Talleres de formación integral.", "Departamen": "Uruguay", "RRSS": "-", "Email": "asocsurco@gmail.com", "lat": -32.475354, "lng": -58.22765 }, "geometry": { "type": "Point", "coordinates": [ -58.227650252153396, -32.47535404117621 ] } },
+    { "type": "Feature", "properties": { "id": 61, "Nombre": "Escuela Privada Especial N°26 El Solar", "Dirección": "Dirección: Alem N° 120, Concepción Del Uruguay, Entre Ríos.", "Teléfono": "(03442) - 422505", "Gestión": "Privada", "Descripcion": "Enseñanzas:Colegios Especiales con Nivel Inicial,Colegios Especiales con Pre Taller,Colegios Primarios Especiales,Colegios Especiales con Talleres", "Departamen": "Uruguay", "RRSS": "-", "Email": "elsolaruruguay@arteibiz.com.ar", "lat": -32.486774, "lng": -58.234219 }, "geometry": { "type": "Point", "coordinates": [ -58.234218768846311, -32.486773669108338 ] } },
+    { "type": "Feature", "properties": { "id": 62, "Nombre": "Escuela Especial Nro 33 Nuevos Horizontes", "Dirección": "Chiesa S/N° , Villaguay, Entre Ríos", "Teléfono": "(03455) - 15452140", "Gestión": "Pública", "Descripcion": "El objetivo de la institución  como centro de recursos múltiples, desde los diversos servicios, es brindar una oferta educativa acorde a las necesidades de todos y cada uno de ellos...", "Departamen": "Villaguay", "RRSS": "https:/www.facebook.com/ESCUELA-DE-EDUCACI%C3%93N-INTEGRAL-N%C2%BA33-192808827447640/", "Email": "escuelaespecial33@hotmail.com", "lat": -31.849422, "lng": -59.021271 }, "geometry": { "type": "Point", "coordinates": [ -59.021270552579573, -31.849422491216863 ] } },
+    { "type": "Feature", "properties": { "id": 63, "Nombre": "Escuela Privada Integral N°7 Dr. Luis Guido", "Dirección": "25 de Mayo, Villaguay, Entre Ríos", "Teléfono": "(03455) - 421076", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Villaguay", "RRSS": "https:/www.facebook.com/Escuela-Privada-de-Educaci%C3%B3n-Integral-N7-Dr-Luis-Guido-689042234638389/", "Email": "epei-nro7@hotmail.com", "lat": -31.86906, "lng": -59.02343 }, "geometry": { "type": "Point", "coordinates": [ -59.023429634990428, -31.869059734773742 ] } },
+    { "type": "Feature", "properties": { "id": 64, "Nombre": "Escuela Especial N°27 Cap. Laboral Elisa A. Ovando.", "Dirección": "Concordia 225,  Villaguay, provincia de Entre Ríos.", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "Es una institución de educación especial con talleres de formación integral de gestión mixta , y de formación laica, se enseña el idioma Inglés en un nivel general.", "Departamen": "Villaguay", "RRSS": "-", "Email": "escuelaovando@live.com", "lat": -31.871924, "lng": -59.030625 }, "geometry": { "type": "Point", "coordinates": [ -59.030624651512454, -31.871924391052687 ] } },
+    { "type": "Feature", "properties": { "id": 65, "Nombre": "Escuela de Equinoterapia La Merced", "Dirección": "H Chiesa, Villaguay, Entre Ríos.", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "La institución ofrece equinoterapia inclusiva gratuita, ya que, se encuentra subvencionada por la municipalidad.", "Departamen": "Villaguay", "RRSS": "https:/www.facebook.com/equinoterapia.villaguay/?ref=page_internal", "Email": "-", "lat": -31.849484, "lng": -59.022387 }, "geometry": { "type": "Point", "coordinates": [ -59.022387014873729, -31.849483783380911 ] } },
+    { "type": "Feature", "properties": { "id": 66, "Nombre": "Escuela Integral Nº 13 Julio B. de Quirós", "Dirección": "Dr. Hernán Cuneo esq. 12 de Octubre, Federal, Entre Ríos.", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "La institución ofrece enseñanza especial en los niveles inicial y primario.", "Departamen": "Federal", "RRSS": "https:/www.facebook.com/groups/352834714810276", "Email": "escuelaespecial13federal@yahoo.com", "lat": -30.946905, "lng": -58.78679 }, "geometry": { "type": "Point", "coordinates": [ -58.786789937951305, -30.946905263708679 ] } },
+    { "type": "Feature", "properties": { "id": 67, "Nombre": "Escuela de Educación Integral Nº 23 Luisa Garnier", "Dirección": "Av. Artigas y Torres Vilches 3190 La Paz, Entre Ríos", "Teléfono": "03437 42-3108", "Gestión": "Pública", "Descripcion": "-", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/Escuela-de-Educaci%C3%B3n-Integral-N%C2%BA-23-Luisa-Garnier-170881007192503/", "Email": "escesp23luisagarnier@hotmail.com", "lat": -30.739171, "lng": -59.621179 }, "geometry": { "type": "Point", "coordinates": [ -59.621179253029865, -30.739171148078821 ] } },
+    { "type": "Feature", "properties": { "id": 68, "Nombre": "Centro Privado Especial N° 6 Mi Estrellita Guia", "Dirección": "9 De Julio 1137, La Paz, entre Ríos", "Teléfono": "03437-421983", "Gestión": "Privada", "Descripcion": "Educación especialEducación especial temprana,Educación especial primaria,Educación especial secundaria", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/escuela.miestrellitaguia", "Email": "esc_especial6@yahoo.com", "lat": -30.743201, "lng": -59.642824 }, "geometry": { "type": "Point", "coordinates": [ -59.642823573620845, -30.743200902414557 ] } },
+    { "type": "Feature", "properties": { "id": 67, "Nombre": "COPNAF", "Dirección": "Belgrano 1143, La Paz. Entre Ríos, Argentina.", "Teléfono": "(03437) 423586", "Gestión": "Pública", "Descripcion": "El COPNAF (Consejo Provincial del Niño, el Adolescente y la Familia) se encuentra descentralizado en los 17 departamentos de la provincia...", "Departamen": "La Paz", "RRSS": "https:/www.entrerios.gov.ar/copnaf/index.php?codigo=2&codsubmenu=81&modulo=&codppal=2", "Email": "copnaflapaz@yahoo.com.ar", "lat": -30.743534, "lng": -59.64246 }, "geometry": { "type": "Point", "coordinates": [ -59.642460363488837, -30.743534154526436 ] } },
+    { "type": "Feature", "properties": { "id": 70, "Nombre": "Centro Ayuda al Niño Aminorado (C.A.N.A)", "Dirección": "9 de julio 1137, la paz 3190, Entre Ríos", "Teléfono": "3437421983", "Gestión": "Pública", "Descripcion": "Atención y ayuda a los discapacitados de la zona.", "Departamen": "La Paz", "RRSS": "-", "Email": "-", "lat": -30.743192, "lng": -59.642899 }, "geometry": { "type": "Point", "coordinates": [ -59.642898762701314, -30.743192350404026 ] } },
+    { "type": "Feature", "properties": { "id": 71, "Nombre": "Dirección de Cultura y Área de Discapacidad", "Dirección": "Echagüe y Moreno.", "Teléfono": "0800-444-5585", "Gestión": "Pública", "Descripcion": "Dicha dirección se encarga de difundir las actividades culturales promovidas por la Dirección de Cultura de La Paz. A continuación, un ejemplo de la Muestra Anual del año 2017.", "Departamen": "La Paz", "RRSS": "https:/www.lapaz.gob.ar/?q=search/node/%C3%84rea%20de%20Discapacidad.", "Email": "direcciondeculturalapaz@gmail.com", "lat": -30.739797, "lng": -59.644885 }, "geometry": { "type": "Point", "coordinates": [ -59.644885200250755, -30.739797492161699 ] } },
+    { "type": "Feature", "properties": { "id": 72, "Nombre": "Escuela Privada De Educación Integral N° 14 Senderos De Luz", "Dirección": "Augusto Timoteo Vandor 255, Bovril, Entre Ríos", "Teléfono": "03438 42-1552", "Gestión": "Privada", "Descripcion": "Un SENDERO...... Que da luz.... Que espera y que da un lugar abriendo miradas, posibilidades y afectos a quien lo necesita.", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/Esc-Priv-Educ-Integral-N14-Senderos-de-Luz-103131694696358/", "Email": "-", "lat": -31.344037, "lng": -59.448256 }, "geometry": { "type": "Point", "coordinates": [ -59.448255624362346, -31.344036948469618 ] } },
+    { "type": "Feature", "properties": { "id": 73, "Nombre": "Escuela de Educación Integral N° 14  María Montessori", "Dirección": "Mariano Moreno y Lisandro de la Torre, Santa Elena, La Paz, Entre Ríos.", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "Título o Certificado seleccionado:Auxiliar en Construcciones.Formación Profesional:Auxiliar Panadero (Auxiliar Panadero y/o Repostero) ", "Departamen": "La Paz", "RRSS": "-", "Email": "-", "lat": -30.949374, "lng": -59.786351 }, "geometry": { "type": "Point", "coordinates": [ -59.786351322171392, -30.949374156747112 ] } },
+    { "type": "Feature", "properties": { "id": 74, "Nombre": "Centro de Integración del Discapacitado CID", "Dirección": "La Bandera N°40, Piedras Blancas, Entre Ríos.", "Teléfono": "03438-495098", "Gestión": "Pública", "Descripcion": "La institución depende de la Secretaría de Desarrollo Social de la Municipalidad de Piedras Blancas. El CID surgió a partir de la necesidad de brindar un ambiente donde...", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/MunicipalidadPiedrasBlancas", "Email": "-", "lat": -31.184709, "lng": -59.950195 }, "geometry": { "type": "Point", "coordinates": [ -59.950194695641059, -31.184709344773992 ] } },
+    { "type": "Feature", "properties": { "id": 75, "Nombre": "Área de Discapacidad de San Gustavo", "Dirección": "Avenida Hernandez y Buenos Aires San Gustavo, Entre Ríos", "Teléfono": "03437 42-4952", "Gestión": "Pública", "Descripcion": "En San Gustavo se trabaja muy bien en discapacidad para que todos puedan sentirse parte, tengan su lugar y la mano tendida ante lo que necesitan...", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/munisangustavo/", "Email": "sangustavoprensa@gmail.com", "lat": -30.690235, "lng": -59.394888 }, "geometry": { "type": "Point", "coordinates": [ -59.394888061847574, -30.690235214433319 ] } },
+    { "type": "Feature", "properties": { "id": 76, "Nombre": "Escuela Especial N°26 PRESBÍTERO LUIS ANTONIO", "Dirección": "Alvear y Santa Fe, Felilciano. Entre Ríos, Argentina", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "Enseñanzas: Colegios con Integración, Estimulación Temprana, Nivel Inicial, Pre Taller, Nivel Primario y Talleres", "Departamen": "La Paz", "RRSS": "-", "Email": "SLILIANAN@HOTMAIL.COM", "lat": -30.386678, "lng": -58.752808 }, "geometry": { "type": "Point", "coordinates": [ -58.752808312228822, -30.386677806180138 ] } },
+    { "type": "Feature", "properties": { "id": 77, "Nombre": "Escuela Privada Integral Melvin Jones", "Dirección": "Cnel. Arredondo 1436, Paraná, Entre Ríos", "Teléfono": "0343 427-9983", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/melvin.jones.399", "Email": "-", "lat": -31.742725, "lng": -60.539748 }, "geometry": { "type": "Point", "coordinates": [ -60.539747517774458, -31.742724687883843 ] } },
+    { "type": "Feature", "properties": { "id": 78, "Nombre": "Centro Educativo Integral y Terapéutico N°7", "Dirección": "Ruta provincial N° 8 Acceso Hernandarias.", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "-", "Departamen": "La Paz", "RRSS": "https:/www.facebook.com/ceit.hernandarias.3", "Email": "-", "lat": -31.237438, "lng": -59.994205 }, "geometry": { "type": "Point", "coordinates": [ -59.994204854163151, -31.237437984740538 ] } },
+    { "type": "Feature", "properties": { "id": 79, "Nombre": "Consultorios Mundo azul", "Dirección": "Dirección:", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "Conformado por: Psicopedagogia, Psicologia, Terapia Ocupacional, Musicoterapia", "Departamen": "Paraná", "RRSS": "https:/instagram.com/consultorios_mundoazul?utm_medium=copy_link", "Email": "consultoriosmundoazul@hotmail.com", "lat": -31.742984, "lng": -60.513946 }, "geometry": { "type": "Point", "coordinates": [ -60.513946175494539, -31.742983741758504 ] } },
+    { "type": "Feature", "properties": { "id": 80, "Nombre": "Centro Panambi", "Dirección": "Cura Alvarez 870, Paraná. Entre Ríos, Argentina.", "Teléfono": "0343 428-0356", "Gestión": "Privada", "Descripcion": "Es un Centro de Intervencion Temprana formado por un equipo de profesionales de diferentes disciplinas que brindan atención a niños de 0 a 6 años.", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/CENTROPANAMBI", "Email": "centropanambi@gmail.com", "lat": -31.74553, "lng": -60.524857 }, "geometry": { "type": "Point", "coordinates": [ -60.524857412316379, -31.745529638225605 ] } },
+    { "type": "Feature", "properties": { "id": 81, "Nombre": "Aloha aprendizaje compartido", "Dirección": "25 de Mayo 760, Paraná, Entre Ríos, Argentina.", "Teléfono": "0343 425-964", "Gestión": "Privada", "Descripcion": "Se dedica a Atención interdisciplinaria. Niños, adolescentes, adultos. Asesoramiento a familias e instituciones escolares. Discapacidad. Talleres recreativos y terapéuticos. Evaluaciones preventivas.", "Departamen": "Paraná", "RRSS": "-", "Email": "aloha_aprendizajecompartido@hotmail.com", "lat": -31.737346, "lng": -60.520285 }, "geometry": { "type": "Point", "coordinates": [ -60.520285330561236, -31.737345722678583 ] } },
+    { "type": "Feature", "properties": { "id": 82, "Nombre": "ACINA: Asociación Centro Integral para el Niño y el Adulto", "Dirección": "Las Tunas 42(Esquina Almafuerte al 1300)-Parana-Entre Rios", "Teléfono": "0343-4241726", "Gestión": "Privada", "Descripcion": "Nuclear a profesionales en las áreas de Psicopedagogía, Fonoaudiología, Kinesiología, Psicología, Servicio Social y otras especialidades afines y complementarias...", "Departamen": "Paraná", "RRSS": "-", "Email": "acina_er@hotmail.com", "lat": -31.752304, "lng": -60.500998 }, "geometry": { "type": "Point", "coordinates": [ -60.500997505591776, -31.75230424323497 ] } },
+    { "type": "Feature", "properties": { "id": 83, "Nombre": "APADEA: Asociación de Padres de Autistas y Trastornos Generales del Desarrollo", "Dirección": "Ramírez 3630 / Moisés Lebenshon y Padre Pascual Uva. Paraná. Entre Ríos.", "Teléfono": "421 962", "Gestión": "Privada", "Descripcion": "Comenzó a formarse en febrero de 2005, integrada por padres, familiares y amigos de personas afectadas por el autismo en la ciudad de Paraná...", "Departamen": "Paraná", "RRSS": "www.apadeaparana.org.ar", "Email": "apadeaparana@yahoo.com.ar", "lat": -31.779863, "lng": -60.532033 }, "geometry": { "type": "Point", "coordinates": [ -60.532032929055802, -31.779862556419666 ] } },
+    { "type": "Feature", "properties": { "id": 84, "Nombre": "APANA: Asociación de Padres y Amigos de Niños y Adolescentes Aminorados.", "Dirección": "25 de Junio 527 Paraná, Entre Ríos / Granja: Av. Lebensohn, Paraná, Entre Ríos.", "Teléfono": "(043)31-6922", "Gestión": "Privada", "Descripcion": "Institución dedicada a la educación y la recreación de niños y adolescentes con discapacidad. El centro de formación brinda herramientas para que las personas puedan tener una salida laboral...", "Departamen": "Paraná", "RRSS": "https:/instagram.com/apanaparana?utm_medium=copy_link", "Email": "apana.parana@gmail.com", "lat": -31.728664, "lng": -60.537136 }, "geometry": { "type": "Point", "coordinates": [ -60.537136262379178, -31.728664449125866 ] } },
+    { "type": "Feature", "properties": { "id": 85, "Nombre": "ASPASID: Asociación Paranaense de Síndrome de Down.", "Dirección": "La Rioja 134, 3114 Paraná, Argentina", "Teléfono": "0343 423-4041", "Gestión": "Privada", "Descripcion": "Fundada el 7 de octubre de 1995 por un grupo de padres en la defensa de los derechos de sus hijos, por una plena integración a todos los ámbitos de la sociedad...", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/aspasidorg", "Email": "aspasid30@gmail.com. / aspasid@arnetbiz.com.ar", "lat": -31.733084, "lng": -60.523712 }, "geometry": { "type": "Point", "coordinates": [ -60.523712175064738, -31.733084420134528 ] } },
+    { "type": "Feature", "properties": { "id": 86, "Nombre": "Escuela de Educación Integral Nº1 H. Keller", "Dirección": "Av. Churruarín 338, E3102 Paraná, Entre Ríos", "Teléfono": "0343 439-0600", "Gestión": "Pública", "Descripcion": "Niveles Educativos: Jardín Maternal Nivel Primario Nivel Secundario", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/Escuela-Helen-Keller-Paran%C3%A1-1909021399176840", "Email": "-", "lat": -31.743512, "lng": -60.513188 }, "geometry": { "type": "Point", "coordinates": [ -60.513187773711685, -31.74351198802524 ] } },
+    { "type": "Feature", "properties": { "id": 87, "Nombre": "Escuela de Educación Integral Nº 3  Carolina Tobar Garcia", "Dirección": "Miguel de Azcuénaga 200, Paraná, Entre Ríos", "Teléfono": "0343 423-3562", "Gestión": "Pública", "Descripcion": "Educación Especial en los Niveles Inicial, Primaria, Ayuda a la Integración Escolar.", "Departamen": "Paraná", "RRSS": "https:/m.facebook.com/Esc3TobarGarcia/?locale2=es_LA", "Email": "escuelacarolinatobargarcia@gmail.com", "lat": -31.722539, "lng": -60.521854 }, "geometry": { "type": "Point", "coordinates": [ -60.521853622742924, -31.722538902987836 ] } },
+    { "type": "Feature", "properties": { "id": 88, "Nombre": "Escuela Educacion Integral Nº 5 Dr. M. Ruiz Moreno", "Dirección": "Don Bosco 749, Paraná, Entre Ríos.", "Teléfono": "-", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/PorUnMundoParaTodos/", "Email": "escespecial@hotmail.com / escintegral5@gmail.com", "lat": -31.734338, "lng": -60.504382 }, "geometry": { "type": "Point", "coordinates": [ -60.50438164125908, -31.734337512586091 ] } },
+    { "type": "Feature", "properties": { "id": 89, "Nombre": "Escuela de Educación Integral nº 7 Edith Fritzgerald", "Dirección": "Gdor. Quirós y Gdor. Domínguez , Paraná, Entre Ríos", "Teléfono": "0221 5564309", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "0113ee0502@gmail.com", "lat": -31.758925, "lng": -60.495915 }, "geometry": { "type": "Point", "coordinates": [ -60.495914709713198, -31.758924924704967 ] } },
+    { "type": "Feature", "properties": { "id": 90, "Nombre": "Escuela de Educación Integral nº 21 Dra. Celia de Montoya", "Dirección": "Facundo y Los Chanas, Paraná, Entre Ríos", "Teléfono": "4270570", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "ESC21CELIAMONTOYA@YAHOO.COM", "lat": -31.740002, "lng": -60.553427 }, "geometry": { "type": "Point", "coordinates": [ -60.553427093116731, -31.740001957373767 ] } },
+    { "type": "Feature", "properties": { "id": 91, "Nombre": "Escuela de Educación Integral Nº1 Zulema Embon", "Dirección": "La Picada, Entre Ríos.", "Teléfono": "0343 499-4024", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/EscuelaZulemaEmbon/", "Email": "-", "lat": -31.735718, "lng": -60.326244 }, "geometry": { "type": "Point", "coordinates": [ -60.326244151969973, -31.735717788787721 ] } },
+    { "type": "Feature", "properties": { "id": 92, "Nombre": "Centro Educativo Terapéutico Nº1", "Dirección": "Suipacha 1870, E3100 Paraná, Entre Ríos", "Teléfono": "0343 424-1166", "Gestión": "Pública", "Descripcion": "Un espacio de formación  no ciclada, donde se piensa un recorrido para cada niño, joven o adulto centrado en sus necesidades, posibilidades e intereses...", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/centro.e.terapeutico", "Email": "-", "lat": -31.731567, "lng": -60.506642 }, "geometry": { "type": "Point", "coordinates": [ -60.50664156522658, -31.731567110327969 ] } },
+    { "type": "Feature", "properties": { "id": 93, "Nombre": "Centro Provincial de Estimulación Temprana.", "Dirección": "Tucumán 333, Paraná, Entre Río", "Teléfono": "0343-4313318", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "-", "lat": -31.726521, "lng": -60.533839 }, "geometry": { "type": "Point", "coordinates": [ -60.533838534185151, -31.726521409325422 ] } },
+    { "type": "Feature", "properties": { "id": 94, "Nombre": "Escuela de Educación Integral Nº20 Luján jaime", "Dirección": "Dr. Elberg 440,  Hasenkamp. Entre Ríos, Argentina.", "Teléfono": "4930448", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/profile.php?id=100009751810016", "Email": "-", "lat": -31.515697, "lng": -59.834194 }, "geometry": { "type": "Point", "coordinates": [ -59.834193507547376, -31.51569666491223 ] } },
+    { "type": "Feature", "properties": { "id": 95, "Nombre": "Escuela de Educación Integral Nº17 Ana M. Cumar de rico", "Dirección": "Dirección: ", "Teléfono": "0343-4949993", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "escuelaespecial17mg@yahoo.com.ar", "lat": -31.66047, "lng": -59.904461 }, "geometry": { "type": "Point", "coordinates": [ -59.904461426814898, -31.660470190509134 ] } },
+    { "type": "Feature", "properties": { "id": 96, "Nombre": "Escuela de Educación Integral Nº27 Antonia L. de Moreyra", "Dirección": "AV. Intendente Aldo Landra 499 , Tabossi. Entre Ríos, Argentina.", "Teléfono": "156296791", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/eei27", "Email": "-", "lat": -31.800612, "lng": -59.935088 }, "geometry": { "type": "Point", "coordinates": [ -59.935087702600903, -31.800611672948801 ] } },
+    { "type": "Feature", "properties": { "id": 97, "Nombre": "Escuela de Educación Integral Nº22 Juan Francisco Seguí", "Dirección": "Bolivar Simon 598, Segui, Entre Río", "Teléfono": "0343 488-048", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "eei22juanfsegui@hotmail.com", "lat": -31.953105, "lng": -60.125866 }, "geometry": { "type": "Point", "coordinates": [ -60.125865809922026, -31.953105389457303 ] } },
+    { "type": "Feature", "properties": { "id": 98, "Nombre": "Escuela de Educación Integral Nº31 Portales De Amor", "Dirección": "Av. 9 de Julio, Cerrito, Entre Ríos", "Teléfono": "0343 489-0607", "Gestión": "Pública", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/eei31portales", "Email": "-", "lat": -31.578572, "lng": -60.075188 }, "geometry": { "type": "Point", "coordinates": [ -60.075187625603647, -31.5785719133745 ] } },
+    { "type": "Feature", "properties": { "id": 99, "Nombre": "Escuela Privada de Educacion Integral Nº9 San Francisco de Asís", "Dirección": "Dra E de Roffo 140, Paraná, Entre Ríos", "Teléfono": "0343 424-3259", "Gestión": "Privada", "Descripcion": "Atención de chicos con discapacidad moderada y severa.", "Departamen": "Paraná", "RRSS": "https:/www.facebook.com/Cooperadora-EPEI-y-Centro-de-Dia-San-Francisco-de-As%C3%ADs-297818221705", "Email": "esc.sanfranciscodeasis04@gmail.com", "lat": -31.749127, "lng": -60.516939 }, "geometry": { "type": "Point", "coordinates": [ -60.516939254931771, -31.749127195035122 ] } },
+    { "type": "Feature", "properties": { "id": 100, "Nombre": "Escuela Privada de Educación Integral Nº16 Pasitos al Futuro", "Dirección": "Rawson 176, Viale, Entre Ríos.", "Teléfono": "-", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "@pasitosalfuturo", "Email": "-", "lat": -31.866842, "lng": -60.003535 }, "geometry": { "type": "Point", "coordinates": [ -60.003534947047967, -31.866842265359345 ] } },
+    { "type": "Feature", "properties": { "id": 101, "Nombre": "Escuela Privada de Educación Integral Nº11 Remedios de Escalada de San Martin", "Dirección": "Roque Saenz Peña 1826, Crespo, Entre Ríos.", "Teléfono": "0343 495-2631", "Gestión": "Privada", "Descripcion": "-", "Departamen": "Paraná", "RRSS": "-", "Email": "-", "lat": -32.035609, "lng": -60.312675 }, "geometry": { "type": "Point", "coordinates": [ -60.312674961573272, -32.035608926761398 ] } }
+    ]
+    }
+
