+++ src/utils/interfaces.ts
@@ -1,0 +1,6 @@
+export type IPathId = 'home' | 'about' | 'places'
+
+export interface IItems {
+    id: IPathId
+    path: string
+}[]
